supports multiple results for the same query

+++ parse.py
@@ -11,7 +11,7 @@ import redis
 class Parse():
 
 	def __init__(self):
-		self.fdate = (date.today()).strftime('%d%m%y')							#keep track of today's date
+		self.fdate = (date.today() - timedelta(1)).strftime('%d%m%y')							#keep track of today's date
 		self.tnum = datetime.today().isoweekday()
 		self.cwd = os.getcwd()
 		self.zipname = self.fdate + ".zip"
@@ -30,11 +30,11 @@ class Parse():
 
 			Only un-comment after deploying the webapp and running the first session.
 		'''
+		print (self.fdate)
+		if (self.fdate == self.last_update() and (self.tnum > 1 and self.tnum < 7)):    	#checks if the database already exists
+			return (self.fromdb())															#return db if it does
 
-		#if (self.fdate == (self.red.lrange("last_updated", 0, 0)[0]).decode('UTF-8')):    	#checks if the database already exists
-		#	return (self.fromdb())															#return db if it does
-
-		#else 
+		#else
 		urlretrieve("http://www.bseindia.com/download/BhavCopy/Equity/EQ" + self.fdate +  "_CSV.ZIP", self.zipname) #retrieve zip file
 
 		zp = zipfile.ZipFile(self.cwd + "/" + self.zipname)
@@ -52,9 +52,11 @@ class Parse():
 			eq = list(csv.DictReader(csvfile))									#parsing csv
 
 		for name in eq:
+			stripped_name = name['SC_NAME'].rstrip()
 			value = {'Name' : name['SC_NAME'].rstrip(), 'Code' : name['SC_CODE'], 'Open' : name['OPEN'], 'Close' : name['CLOSE'], 'High' : name['HIGH'], 'Low' : name['LOW']}
-			self.red.hmset(name['SC_NAME'].rstrip(), value)						#set dict to an hash in redis
-			self.red.rpush("insertion_order", name['SC_NAME'].rstrip())			#keep track of names in a different list
+			self.red.hmset(stripped_name, value)								#set dict to an hash in redis
+			self.red.rpush("insertion_order", stripped_name)					#keep track of names in a different list
+			self.red.zadd("search_bd", stripped_name, 0)						#keep track of names via sorted list for better search functionality
 
 		self.red.lpush("last_updated", self.fdate)								#keep track of last update date
 		return (self.fromdb())
@@ -71,4 +73,14 @@ class Parse():
 		return ((self.red.lrange("last_updated", 0, 0)[0]).decode('UTF-8'))		#return the last update date
 
 	def search(self, q):														#searching for a particular entry and returing the same if it exists
-			return (self.red.hgetall(q))
+		matches = self.red.zrangebylex("search_bd", "[" + q, "[" + q + "\xff")	#fetch and store the matched entries
+		reg = []
+		if len(matches) != 0:													#get all the details and parse
+			if len(matches) < 10:
+				for match in matches:
+					reg.append(self.red.hgetall(match))
+			else:
+				for i in range (10):
+					reg.append(self.red.hgetall(matches[i]))
+
+		return (reg)
